Make filters less dependend on each other (potentially breaks your filters).
Previously when you mapped, e.g., a locale filter and a pagination filter, routing_filter would run the filters in this order for both recognition and generation. That means that both filters need to know more about potential changes of the other filter in around_recognition or around_generation respectively. 

This change makes it so that filters are run in the registration order for #around_recognize (i.e. input comes in from the outer world), but are run in the reverse order for #around_generate (i.e. output comes from your app).